requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openai>=1.50.0
python-dotenv>=1.0.0
apscheduler>=3.10.4
httpx>=0.27.0
aiofiles>=24.1.0
psycopg2-binary>=2.9.9
